changes comment_count to INT

+++ app/model/model.js
@@ -40,7 +40,7 @@ exports.selectArticles = () => {
         articles.created_at,
         articles.votes,
         articles.article_img_url,
-        COUNT(comments.comment_id) AS comment_count
+        COUNT(comments.comment_id)::INT AS comment_count
         FROM articles
         LEFT JOIN 
         comments ON articles.article_id = comments.article_id
